requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
async-timeout==5.0.1
asyncio-dgram==2.2.0
attrs==25.3.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chess==1.11.2
cssselect2==0.8.0
defusedxml==0.7.1
discord.py==2.5.2
discordLevelingSystem==1.2.1
dnspython==2.7.0
fake-useragent==2.2.0
frozenlist==1.6.0
idna==3.10
mcstatus==11.1.1
multidict==6.4.3
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
PyNaCl==1.5.0
rcon==2.1.1
requests==2.32.3
tinycss2==1.4.0
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
yarl==1.20.0
yt-dlp==2025.4.30
